+++ src/main/java/org/example/spring/beans/factory/support/CglibInstantiationStrategy.java
@@ -1,0 +1,19 @@
+package org.example.spring.beans.factory.support;
+
+import net.sf.cglib.proxy.Enhancer;
+import net.sf.cglib.proxy.MethodInterceptor;
+import org.example.spring.beans.exception.BeanException;
+
+/**
+ * @Author Roc
+ * @Date 2024/11/20 16:06
+ */
+public class CglibInstantiationStrategy implements InstantiationStrategy {
+    @Override
+    public Object instantiate(BeanDefinition beanDefinition) throws BeanException {
+        Enhancer enhancer = new Enhancer();
+        enhancer.setSuperclass(beanDefinition.getBeanClass());
+        enhancer.setCallback((MethodInterceptor) (obj, method, argsTemp, proxy) -> proxy.invokeSuper(obj, argsTemp));
+        return enhancer.create();
+    }
+}