+++ changelog.md
@@ -91,3 +91,18 @@ public class HelloService {
     }
 }
 ```
+
+## Bean实例化策略InstantiationStrategy
+
+
+
+> 代码分支：3-instantiation-strategy
+
+现在bean是在AbstractAutowireCapableBeanFactory.doCreateBean方法中用beanClass.newInstance()来实例化，仅适用于bean有无参构造函数的情况。
+
+[![img](.\image\3\PixPin_2024-11-20_16-20-45.png)](https://github.com/DerekYRC/mini-spring/blob/main/assets/instantiation-strategy.png)
+
+针对bean的实例化，抽象出一个实例化策略的接口InstantiationStrategy，有两个实现类：
+
+- SimpleInstantiationStrategy，使用bean的构造函数来实例化
+- CglibInstantiationStrategy，使用CGLIB动态生成子类